STCOM-1099 Added tests for DateRangeWrapper Final Form
